+++ src/test/compile-fail/generic-type-more-params-with-defaults.rs
@@ -8,13 +8,11 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-#![feature(default_type_params)]
-
 struct Heap;
 
 struct Vec<T, A = Heap>;
 
 fn main() {
-    let _: Vec<int, Heap, bool>;
+    let _: Vec<isize, Heap, bool>;
     //~^ ERROR wrong number of type arguments: expected at most 2, found 3
 }
